Unmount flash disk so it can be safely removed
Developed for Xubuntu operating system.

Software Python version 3.7 or higher is required to be installed, because the argument 'capture_output' of the method 'subprocess.run' was added in python 3.7 .
Software Zenity is required to be installed.
Software lsblk is required to be installed.
Software udisksctl is required to be installed.

Example of usage on linux command line: python3.7 unmount_flash_disk.py;

Optional argument --help or -h can be given.

Prepare the application for all possible situations and errors.